chore: restruct workspace controller code - part 4
